refactor: migrate layout components to typescript
