--- src/components/layout/index.js
@@ -1,1 +1,0 @@
-export { default as Nav } from "./Nav";
